Move CandidateKeywordsTrait.php into app/Traits
