+++ app/Traits/CandidateKeywordsTrait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformers;
+namespace App\Traits;
 
 trait CandidateKeywordsTrait
 {
